requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2.10
djangorestframework==3.16.1
environs==14.5.0
idna==3.11
marshmallow==4.2.0
psycopg2==2.9.11
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
